+++ crate/server/src/config/params/server_params.rs
@@ -73,7 +73,7 @@ pub struct ServerParams {
     ///
     /// It should contain the external URL of this server as configured in Google Workspace client side encryption settings
     /// For instance, if this server is running on domain cse.my_domain.com,
-    /// the URL should be something like https://cse.my_domain.com/google_cse
+    /// the URL should be something like <https://cse.my_domain.com/google_cse>
     pub google_cse_kacls_url: Option<String>,
 }