+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [X.Y.Z] - 2023-XX-YY
+
+### Features
+
+- adds support for Google Workspace Client Side Encryption. Single key for now.
+- ckms: adds support for login to the Identity Provider and get an access token
+
 ## [4.9.1] - 2023-11-14
 
 ### Bug Fixes